Move refer-friends helpers into dedicated file per https://github.com/DoSomething/phoenix-next/pull/2168#discussion_r430559146
